requirements.txt
fastapi
uvicorn
neo4j
faiss-cpu
sentence-transformers
numpy
python-dotenv
requests
pydantic
streamlit
streamlit-agraph
